Split up code to catch ImportError better
