+++ ynr/apps/sopn_parsing/tests/test_helpers.py
@@ -3,16 +3,21 @@ from unittest import skipIf
 
 from django.test import TestCase
 
-from sopn_parsing.helpers import clean_text, SOPNDocument
+from sopn_parsing.helpers.text_helpers import clean_text
 from sopn_parsing.tests import should_skip_pdf_tests
 
+try:
+    from sopn_parsing.helpers.pdf_helpers import SOPNDocument
+except ImportError:
+    pass
 
+
+@skipIf(should_skip_pdf_tests(), "Required PDF libs not installed")
 class TestSOPNHelpers(TestCase):
     def test_clean_text(self):
         text = "\n C andidates (Namés)"
         self.assertEqual(clean_text(text), "candidates")
 
-    @skipIf(should_skip_pdf_tests(), "Required PDF libs not installed")
     def test_sopn_document(self):
         example_doc_path = abspath(
             join(dirname(__file__), "data/sopn-berkeley-vale.pdf")
@@ -70,7 +75,6 @@ class TestSOPNHelpers(TestCase):
             doc.get_pages_by_ward_name("berkeley")[0].page_number, 1
         )
 
-    @skipIf(should_skip_pdf_tests(), "Required PDF libs not installed")
     def test_multipage_doc(self):
         example_doc_path = abspath(
             join(dirname(__file__), "data/NI-Assembly-Election-2016.pdf")
